+++ R/zzz.R
@@ -1,9 +1,8 @@
-#' @keywords internal
 .onAttach <- function(...) {
-  attached <- coursekata_attach()
+  attached <- coursekata_attach(quietly = getOption("coursekata.quiet", FALSE) || quickstart())
   coursekata_load_theme()
   if (!quickstart()) {
-    coursekata_attach_message(attached)
+    rlang::inform(coursekata_attach_message(attached), class = "packageStartupMessage")
   }
 }